+++ process3_wipsvg-xml_MONO.sh
@@ -1,5 +1,9 @@
 #! /bin/bash
 
+# Check dependencies
+command -v inkscape >/dev/null 2>&1 || { echo >&2 "I require inkscape but it's not installed. Aborting."; exit 1; }
+command -v scour >/dev/null 2>&1 || { echo >&2 "I require scour but it's not installed. Aborting."; exit 1; }
+
 # Delete current icons to re-monochrome (yes, despite how long this takes, its a necessary evil)
 rm -rf app/src/monochromatic/res/drawable-mdpi
 rm -rf app/src/monochromatic/res/drawable-hdpi
@@ -27,21 +31,16 @@ fi
 # get from icons/monochromatic
 for SVG in icons/monochromatic/*.svg
 do
-  if [[ -f "${SVG}" ]]; then
-    N=$(basename ${SVG} .svg)
-    inkscape -w 48 -h 48 -o \
-        app/src/monochromatic/res/drawable-mdpi/${N}.png ${SVG}
-    inkscape -w 72 -h 72 -o \
-        app/src/monochromatic/res/drawable-hdpi/${N}.png ${SVG}
-    inkscape -w 96 -h 96 -o \
-        app/src/monochromatic/res/drawable-xhdpi/${N}.png ${SVG}
-    inkscape -w 144 -h 144 -o \
-        app/src/monochromatic/res/drawable-xxhdpi/${N}.png ${SVG}
-    inkscape -w 192 -h 192 -o \
-        app/src/monochromatic/res/drawable-xxxhdpi/${N}.png ${SVG}
+    if [[ -f "${SVG}" ]]; then
+        N=$(basename ${SVG} .svg)
+        inkscape -w 48 -h 48 -o app/src/monochromatic/res/drawable-mdpi/${N}.png ${SVG}
+        inkscape -w 72 -h 72 -o app/src/monochromatic/res/drawable-hdpi/${N}.png ${SVG}
+        inkscape -w 96 -h 96 -o app/src/monochromatic/res/drawable-xhdpi/${N}.png ${SVG}
+        inkscape -w 144 -h 144 -o app/src/monochromatic/res/drawable-xxhdpi/${N}.png ${SVG}
+        inkscape -w 192 -h 192 -o app/src/monochromatic/res/drawable-xxxhdpi/${N}.png ${SVG}
 
-    cp ${SVG} ${SVG}.tmp
-    scour --remove-descriptive-elements --enable-id-stripping --enable-viewboxing --enable-comment-stripping --nindent=4 -i ${SVG}.tmp -o ${SVG}
-    rm ${SVG}.tmp
-  fi
+        cp ${SVG} ${SVG}.tmp
+        scour --remove-descriptive-elements --enable-id-stripping --enable-viewboxing --enable-comment-stripping --nindent=4 -i ${SVG}.tmp -o ${SVG}
+        rm ${SVG}.tmp
+    fi
 done
